refactor(biz): relocate functions closer to their callers
Why this change was needed:
Functions were placed in common.go but only used by single modules,
creating unnecessary indirection. Moving them closer to their callers
improves code cohesion and makes dependencies clearer.

What changed:
- Moved buildCommonDbQuery from common.go to mcp_helpers.go
- Moved FormatResourceIDNameList from common.go to publish.go
- Removed unused gorm import from common.go

Problem solved:
common.go was acting as a catch-all module. By relocating functions
to their actual usage sites, the codebase structure becomes more
intentional and easier to navigate.
