+++ src/apiserver/pkg/biz/common.go
@@ -31,7 +31,6 @@ import (
 	"gorm.io/datatypes"
 	"gorm.io/gen"
 	"gorm.io/gen/field"
-	"gorm.io/gorm"
 
 	"github.com/tidwall/sjson"
 
@@ -120,16 +119,6 @@ func (l Labels) Value() (driver.Value, error) {
 		return nil, nil
 	}
 	return json.Marshal(l)
-}
-
-// buildCommonDbQuery 获取通用查询
-// 该查询会根据网关 ID 进行过滤，确保只能查询当前网关下的资源
-func buildCommonDbQuery(
-	ctx context.Context,
-	resourceType constant.APISIXResource,
-) *gorm.DB {
-	return database.Client().WithContext(ctx).Table(resourceTableMap[resourceType]).Where(
-		"gateway_id = ?", ginx.GetGatewayInfoFromContext(ctx).ID)
 }
 
 // BatchDeleteResourceByIDs 事务批量删除资源
@@ -939,61 +928,4 @@ func BuildConfigRawForValidation(
 	}
 
 	return configRawForValidation
-}
-
-// FormatResourceIDNameList 格式化资源 ID 和名称列表
-func FormatResourceIDNameList(resources any, resourceType constant.APISIXResource) []string {
-	switch resourceType {
-	case constant.Route:
-		routes := resources.([]*model.Route) //nolint:forcetypeassert
-		routeDetails := make([]string, 0, len(routes))
-		for _, route := range routes {
-			routeDetails = append(
-				routeDetails,
-				fmt.Sprintf("%s(%s)", route.ID, route.GetName(resourceType)),
-			)
-		}
-		return routeDetails
-	case constant.Upstream:
-		upstreams := resources.([]*model.Upstream) //nolint:forcetypeassert
-		upstreamDetails := make([]string, 0, len(upstreams))
-		for _, upstream := range upstreams {
-			upstreamDetails = append(
-				upstreamDetails,
-				fmt.Sprintf("%s(%s)", upstream.ID, upstream.GetName(resourceType)),
-			)
-		}
-		return upstreamDetails
-	case constant.Consumer:
-		consumers := resources.([]*model.Consumer) //nolint:forcetypeassert
-		consumerDetails := make([]string, 0, len(consumers))
-		for _, consumer := range consumers {
-			consumerDetails = append(
-				consumerDetails,
-				fmt.Sprintf("%s(%s)", consumer.ID, consumer.GetName(resourceType)),
-			)
-		}
-		return consumerDetails
-	case constant.Service:
-		services := resources.([]*model.Service) //nolint:forcetypeassert
-		serviceDetails := make([]string, 0, len(services))
-		for _, service := range services {
-			serviceDetails = append(
-				serviceDetails,
-				fmt.Sprintf("%s(%s)", service.ID, service.GetName(resourceType)),
-			)
-		}
-		return serviceDetails
-	case constant.StreamRoute:
-		streamRoutes := resources.([]*model.StreamRoute) //nolint:forcetypeassert
-		streamRouteDetails := make([]string, 0, len(streamRoutes))
-		for _, streamRoute := range streamRoutes {
-			streamRouteDetails = append(
-				streamRouteDetails,
-				fmt.Sprintf("%s(%s)", streamRoute.ID, streamRoute.GetName(resourceType)),
-			)
-		}
-		return streamRouteDetails
-	}
-	return nil
 }
